+++ Cipher/algorithms/random/xoroshiro256starstar.js
@@ -245,7 +245,7 @@
      *
      * Algorithm from https://prng.di.unimi.it/xoroshiro256starstar.c:
      * 1. result = rotl(s[1] * 5, 7) * 9  (**starstar** scrambler)
-     * 2. t = s[1] << 17
+     * 2. t = s[1] shl 17
      * 3. s[2] ^= s[0]
      * 4. s[3] ^= s[1]
      * 5. s[1] ^= s[2]
@@ -275,8 +275,8 @@
       const result = OpCodes.ToQWord(rotated * 9n);
 
       // State update
-      // t = s[1] << 17
-      const t = OpCodes.ToQWord(OpCodes.ShiftLn(s1, 17));
+      // t = s[1] shl 17
+      const t = OpCodes.ToQWord(OpCodes.ShiftLn(s1, 17n));
 
       // s[2] ^= s[0]
       // s[3] ^= s[1]
@@ -323,7 +323,7 @@
 
         // Extract bytes in little-endian order
         for (let i = 0; i < 8 && bytesGenerated < length; ++i) {
-          const shifted = OpCodes.ShiftRn(value64, i * 8);
+          const shifted = OpCodes.ShiftRn(value64, BigInt(i * 8));
           const byteVal = Number(OpCodes.AndN(shifted, 0xFFn));
           output.push(byteVal);
           ++bytesGenerated;
@@ -390,7 +390,7 @@
 
       for (let i = 0; i < JUMP.length; ++i) {
         for (let b = 0; b < 64; ++b) {
-          const mask = OpCodes.ShiftLn(1n, b);
+          const mask = OpCodes.ShiftLn(1n, BigInt(b));
           if (OpCodes.AndN(JUMP[i], mask) !== 0n) {
             s0 = OpCodes.XorN(s0, this._s0);
             s1 = OpCodes.XorN(s1, this._s1);
@@ -428,7 +428,7 @@
 
       for (let i = 0; i < LONG_JUMP.length; ++i) {
         for (let b = 0; b < 64; ++b) {
-          const mask = OpCodes.ShiftLn(1n, b);
+          const mask = OpCodes.ShiftLn(1n, BigInt(b));
           if (OpCodes.AndN(LONG_JUMP[i], mask) !== 0n) {
             s0 = OpCodes.XorN(s0, this._s0);
             s1 = OpCodes.XorN(s1, this._s1);
